+++ src/ledger/OfferHelper.h
@@ -40,6 +40,8 @@ namespace stellar
         EntryFrame::pointer fromXDR(LedgerEntry const& from) override;
         uint64_t countObjects(soci::session& sess) override;
 
+        bool exists(Database& db, AssetCode const& base, AssetCode const& quote, uint64_t* orderBookID);
+
         OfferFrame::pointer loadOffer(AccountID const& accountID, uint64_t offerID,
                           Database& db, LedgerDelta* delta = nullptr);