+++ src/ledger/SaleHelper.h
@@ -39,6 +39,7 @@ namespace stellar
         uint64_t countObjects(soci::session& sess) override;
 
         bool exists(Database& db, uint64_t saleID);
+        bool exists(Database& db, AssetCode const& base, AssetCode const& quote);
 
         SaleFrame::pointer loadSale(uint64_t saleID, AccountID ownerID, Database& db, LedgerDelta* delta = nullptr);
         SaleFrame::pointer loadSale(uint64_t saleID, Database& db, LedgerDelta* delta = nullptr);